+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Use collection version in User-Agent header instead of Ansible version.
+
 ## [0.8.0] - 2025-08-18
 
 ### Added